+++ libs/api/workflows/src/core/job-execution-outcome.ts
