+++ libs/api/workflows/src/temporal/workflows/deadline.ts
@@ -1,0 +1,7 @@
+export function deadlineReached(deadline: number | undefined): boolean {
+  return deadline !== undefined && Date.now() >= deadline;
+}
+
+export function remainingMs(deadline: number | undefined): number | undefined {
+  return deadline === undefined ? undefined : Math.max(0, deadline - Date.now());
+}
